+++ src/Pickles/Pickles.TestFrameworks.UnitTests/StandardTestSuite.cs
@@ -245,6 +245,31 @@ namespace PicklesDoc.Pickles.TestFrameworks.UnitTests
             Check.That(actualResult).IsEqualTo(TestResult.Failed);
         }
 
+        public void ThenCanReadResultOfScenarioWithSpecialCharacters()
+        {
+            var results = ParseResultsFile();
+
+            var feature = new Feature { Name = "Scenarios With Special Characters" };
+
+            var scenario = new Scenario{ Name = "This is a scenario with parentheses, hyphen and comma (10-20, 30-40)", Feature = feature };
+
+            var actualResult = results.GetScenarioResult(scenario);
+            Check.That(actualResult).IsEqualTo(TestResult.Passed);
+        }
+
+        public void ThenCanReadResultOfScenarioOutlineWithSpecialCharacters()
+        {
+            var results = ParseResultsFile();
+
+            var feature = new Feature { Name = "Scenarios With Special Characters" };
+
+            var scenarioOutline = new ScenarioOutline { Name = "This is a scenario outline with parentheses, hyphen and comma (10-20, 30-40)", Feature = feature };
+            
+            var actualResult = results.GetExampleResult(scenarioOutline, new string[] { "pass_1" });
+
+            Check.That(actualResult).IsEqualTo(TestResult.Passed);
+        }
+
         private Feature AdditionFeature()
         {
             return new Feature { Name = "Addition" };