+++ src/Pickles/Pickles.TestFrameworks/NUnit/NUnitExampleSignatureBuilder.cs
@@ -34,17 +34,12 @@ namespace PicklesDoc.Pickles.TestFrameworks.NUnit
         {
             var stringBuilder = new StringBuilder();
 
-            var name = scenarioOutline.Name.ToLowerInvariant();
-            name = PunctuationCharactersRegex.Replace(name, "_");
-            name = NonIdentifierCharacterRegex.Replace(name, string.Empty);
+            var name = SpecFlowNameMapping.Build(scenarioOutline.Name.ToLowerInvariant());
             stringBuilder.Append(name).Append("\\(");
 
             foreach (string value in row)
             {
-                stringBuilder.AppendFormat("\"{0}\",", value.ToLowerInvariant()
-                    .Replace(@"\", string.Empty).Replace(@"$", @"\$")
-                    .Replace(@"(", @"\(")
-                    .Replace(@")", @"\)"));
+                stringBuilder.AppendFormat("\"{0}\",", Regex.Escape(value.ToLowerInvariant()));
             }
 
             stringBuilder.Remove(stringBuilder.Length - 1, 1);